Start abstracting platforms in ykcompile.
This change lays the foundations for platform abstration:

 - Moves the (currently x86_64-centric) `TraceCompiler` struct into
   arch/x86_64/mod.rs.

 - Leaves all generic structs in the top-level to be shared across
   future backends.

 - For now, hard-wires use of the x86_64 backend.

In the long run, we should slice up `TraceCompiler` into some kind of a
modular API (statically dispatched trait?), but this change does not
address this.
