+++ internal_ws/ykcompile/src/arch/mod.rs
@@ -1,0 +1,7 @@
+//! Platform support for the Yorick TIR trace compiler.
+
+#[cfg(not(all(target_arch = "x86_64", target_os = "linux")))]
+compile_error!("Currently only linux x86_64 is supported.");
+
+#[cfg(all(target_arch = "x86_64", target_os = "linux"))]
+pub(crate) mod x86_64;
